# openapi.yaml
openapi: 3.0.3
info:
  title: cerbix API
  version: 0.19.0
  description: >
    Internal uptime & SLA monitoring API. Authentication is via a server-side
    session cookie established by an OIDC provider (`/auth/login` → `/auth/callback`)
    or local login (`/auth/local/login`). All `/api/v1/*` endpoints require a
    valid session and enforce tenant isolation: callers only see and mutate the
    organizations/projects they are authorized for (non-members receive 404;
    insufficient in-org role receives 403).
servers:
  - url: /
    description: Same-origin (SPA is served by the API binary)
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
  /auth/local/login:
    post:
      tags: [auth]
      summary: Local username/password login (issues a session cookie)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated; sets the session cookie.
          content:
            application/json:
              schema: { type: object, properties: { ok: { type: boolean } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/local/reset/request:
    post:
      tags: [auth]
      summary: Request a password-reset link (local accounts)
      security: []
      description: >
        Emails a one-time reset link if the address belongs to a local account.
        Always returns 200 with a generic body so it cannot enumerate accounts.
        Requires outbound email (mail) to be configured; see /auth/config
        password_reset.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: Accepted (whether or not the address exists).
          content:
            application/json:
              schema: { type: object, properties: { ok: { type: boolean } } }
        '429': { description: Too many requests. }
  /auth/local/reset/confirm:
    post:
      tags: [auth]
      summary: Complete a password reset with a token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token: { type: string, description: The token from the reset link. }
                new_password: { type: string, format: password }
      responses:
        '204': { description: Password changed. }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /auth/login:
    get:
      tags: [auth]
      summary: Begin OIDC login (redirects to the configured identity provider)
      security: []
      parameters:
        - in: query
          name: redirect
          schema: { type: string }
          description: Local path to return to after login.
      responses:
        '302': { description: Redirect to the identity provider. }
  /auth/callback:
    get:
      tags: [auth]
      summary: OIDC redirect URI (exchanges code, sets the session cookie)
      security: []
      responses:
        '302': { description: Redirect to the post-login path. }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cerbix_session
    bearerAuth:
      type: http
      scheme: bearer
      description: "Machine auth via Authorization: Bearer. Either a cerbix service-account token (cbx_...) or an OIDC client-credentials access token (a realm-issued JWT, when OIDC is enabled); the JWT subject is JIT-provisioned and access is membership-gated."
...
  responses:
    Unauthorized:
      description: No or invalid session.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, description: Email of the local account. }
        password: { type: string, format: password }
        totp: { type: string, description: 'TOTP code or a recovery code, required when 2FA is enabled.' }
...